requirements.txt
Flask
Flask-BasicAuth
email-validator

google-cloud-datastore
google-cloud-storage